+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Version 6.1.1
 
+## Merged from Version 6.0
+
+* Refactor ecchronos-binary Test Infrastructure - Issue #1093
 
 ## Version 6.1.0
 
@@ -10,6 +13,7 @@
 
 ## Version 6.0.8
 
+* Refactor ecchronos-binary Test Infrastructure - Issue #1093
 
 ## Version 6.0.7